Do QA for DogWalkers and Trips + add tripId

+++ app/src/main/java/com/example/nofarcohenzedek/dogo/Model/DogWalker.java
@@ -63,7 +63,7 @@ public class DogWalker extends User{
 
         if(this.comments.size() != 0)
         {
-            this.averageRating = sum / this.comments.size();
+            this.averageRating = (double)sum / this.comments.size();
         }
     }
